:tada: moved categories manager to component
